+++ examples/regression/2-advanced-analysis/plot_timeseries_enbpi.py
@@ -2,28 +2,27 @@
 ==================================================================
 Estimating prediction intervals of time series forecast with EnbPI
 ==================================================================
+
 This example uses
 :class:`mapie.time_series_regression.MapieTimeSeriesRegressor` to estimate
-prediction intervals associated with time series forecast. It follows [6] and
-an alternative expermimental implemetation inspired from [2]
+prediction intervals associated with time series forecast. It follows [6].
 
 We use here the Victoria electricity demand dataset used in the book
 "Forecasting: Principles and Practice" by R. J. Hyndman and G. Athanasopoulos.
 The electricity demand features daily and weekly seasonalities and is impacted
 by the temperature, considered here as a exogeneous variable.
 
-A Random Forest model is aloready fitted on data. The hyper-parameters are
+A Random Forest model is already fitted on data. The hyper-parameters are
 optimized with a :class:`sklearn.model_selection.RandomizedSearchCV` using a
 sequential :class:`sklearn.model_selection.TimeSeriesSplit` cross validation,
 in which the training set is prior to the validation set.
 The best model is then feeded into
 :class:`mapie.time_series_regression.MapieTimeSeriesRegressor` to estimate the
 associated prediction intervals. We compare four approaches: with or without
-``partial_fit`` called at every step, and following [6] or a approach inspired
-from [2]. It appears that the approach inspired from [2] and ``partial_fit``
-offer higher coverage, but with higher width of PIs and are much slower.
+``partial_fit`` called at every step, and following [6]. It appears that
+``partial_fit`` offer higher coverage, but with higher width of PIs and is much
+slower.
 """
-import warnings
 
 import numpy as np
 import pandas as pd
@@ -32,11 +31,12 @@ from scipy.stats import randint
 from sklearn.ensemble import RandomForestRegressor
 from sklearn.model_selection import RandomizedSearchCV, TimeSeriesSplit
 
-from mapie.metrics import regression_coverage_score
+from mapie.metrics import (
+    regression_coverage_score,
+    regression_mean_width_score,
+)
 from mapie.subsample import BlockBootstrap
 from mapie.time_series_regression import MapieTimeSeriesRegressor
-
-warnings.simplefilter("ignore")
 
 # Load input data and feature engineering
 demand_df = pd.read_csv(
@@ -66,8 +66,8 @@ y_train = demand_train.loc[X_train.index, "Demand"]
 X_test = demand_test.loc[:, features]
 y_test = demand_test["Demand"]
 
-model_params_fit_not_done = False
-if model_params_fit_not_done:
+perform_hyperparameters_search = False
+if perform_hyperparameters_search:
     # CV parameter search
     n_iter = 100
     n_splits = 5
@@ -90,104 +90,69 @@ if model_params_fit_not_done:
 else:
     # Model: Random Forest previously optimized with a cross-validation
     model = RandomForestRegressor(
-        max_depth=10, n_estimators=50, random_state=59)
+        max_depth=10, n_estimators=50, random_state=59
+    )
 
 # Estimate prediction intervals on test set with best estimator
 alpha = 0.05
-cv_MapieTimeSeries = BlockBootstrap(
+cv_mapietimeseries = BlockBootstrap(
     n_resamplings=100, length=48, overlapping=True, random_state=59
 )
 
-mapie_plus = MapieTimeSeriesRegressor(
-    model, method="plus", cv=cv_MapieTimeSeries, agg_function="mean", n_jobs=-1
-)
 mapie_enpbi = MapieTimeSeriesRegressor(
-    model, method="plus", cv=cv_MapieTimeSeries, agg_function="mean", n_jobs=-1
+    model,
+    method="enbpi",
+    cv=cv_mapietimeseries,
+    agg_function="mean",
+    n_jobs=-1,
 )
-
-gap = 1
 
 print("EnbPI, with no partial_fit, width optimization")
 mapie_enpbi = mapie_enpbi.fit(X_train, y_train)
 y_pred_npfit_enbpi, y_pis_npfit_enbpi = mapie_enpbi.predict(
-    X_test, alpha=alpha, ensemble=True, beta_optimize=True
+    X_test, alpha=alpha, ensemble=True, optimize_beta=True
 )
 coverage_npfit_enbpi = regression_coverage_score(
     y_test, y_pis_npfit_enbpi[:, 0, 0], y_pis_npfit_enbpi[:, 1, 0]
 )
-width_npfit_enbpi = (
-    y_pis_npfit_enbpi[:, 1, 0] - y_pis_npfit_enbpi[:, 0, 0]
-).mean()
+
+width_npfit_enbpi = regression_mean_width_score(
+    y_pis_npfit_enbpi[:, 1, 0], y_pis_npfit_enbpi[:, 0, 0]
+)
 
 print("EnbPI with partial_fit, width optimization")
 mapie_enpbi = mapie_enpbi.fit(X_train, y_train)
 y_pred_pfit_enbpi = np.zeros(y_pred_npfit_enbpi.shape)
 y_pis_pfit_enbpi = np.zeros(y_pis_npfit_enbpi.shape)
 
-
-y_pred_pfit_enbpi[:gap], y_pis_pfit_enbpi[:gap, :, :] = mapie_enpbi.predict(
-    X_test.iloc[:gap, :], alpha=alpha, ensemble=True, beta_optimize=True
+step_size = 1
+(
+    y_pred_pfit_enbpi[:step_size],
+    y_pis_pfit_enbpi[:step_size, :, :],
+) = mapie_enpbi.predict(
+    X_test.iloc[:step_size, :], alpha=alpha, ensemble=True, optimize_beta=True
 )
 
-for step in range(gap, len(X_test), gap):
+for step in range(step_size, len(X_test), step_size):
     mapie_enpbi.partial_fit(
-        X_test.iloc[(step - gap):step, :],
-        y_test.iloc[(step - gap):step],
+        X_test.iloc[(step - step_size) : step, :],
+        y_test.iloc[(step - step_size) : step],
     )
     (
-        y_pred_pfit_enbpi[step:step + gap],
-        y_pis_pfit_enbpi[step:step + gap, :, :],
+        y_pred_pfit_enbpi[step : step + step_size],
+        y_pis_pfit_enbpi[step : step + step_size, :, :],
     ) = mapie_enpbi.predict(
-        X_test.iloc[step:(step + gap), :],
+        X_test.iloc[step : (step + step_size), :],
         alpha=alpha,
         ensemble=True,
-        beta_optimize=True,
+        optimize_beta=True,
     )
 coverage_pfit_enbpi = regression_coverage_score(
     y_test, y_pis_pfit_enbpi[:, 0, 0], y_pis_pfit_enbpi[:, 1, 0]
 )
-width_pfit_enbpi = (
-    y_pis_pfit_enbpi[:, 1, 0] - y_pis_pfit_enbpi[:, 0, 0]
-).mean()
-
-print("Plus, with partial_fit, width optimization")
-mapie_plus = mapie_plus.fit(X_train, y_train)
-y_pred_pfit_plus = np.zeros(y_pred_npfit_enbpi.shape)
-y_pis_pfit_plus = np.zeros(y_pis_npfit_enbpi.shape)
-(y_pred_pfit_plus[:gap], y_pis_pfit_plus[:gap, :, :],) = mapie_plus.predict(
-    X_test.iloc[:gap, :],
-    alpha=alpha,
-    beta_optimize=True,
+width_pfit_enbpi = regression_mean_width_score(
+    y_pis_pfit_enbpi[:, 1, 0], y_pis_pfit_enbpi[:, 0, 0]
 )
-for step in range(gap, len(X_test), gap):
-    mapie_plus.partial_fit(
-        X_test.iloc[step - gap:step, :],
-        y_test.iloc[step - gap:step],
-    )
-    (
-        y_pred_pfit_plus[step:step + gap],
-        y_pis_pfit_plus[step:step + gap, :, :],
-    ) = mapie_plus.predict(
-        X_test.iloc[step:step + gap, :],
-        alpha=alpha,
-        ensemble=True,
-        beta_optimize=True,
-    )
-
-coverage_pfit_plus = regression_coverage_score(
-    y_test, y_pis_pfit_plus[:, 0, 0], y_pis_pfit_plus[:, 1, 0]
-)
-width_pfit_plus = (y_pis_pfit_plus[:, 1, 0] - y_pis_pfit_plus[:, 0, 0]).mean()
-
-print("Plus, with NO partial_fit, MapieRegressor_Like no")
-mapie_plus = mapie_plus.fit(X_train, y_train)
-y_pred_pfit_MR, y_pis_pfit_MR = mapie_plus.predict(
-    X_test, alpha=alpha, ensemble=True
-)
-coverage_pfit_MR = regression_coverage_score(
-    y_test, y_pis_pfit_MR[:, 0, 0], y_pis_pfit_MR[:, 1, 0]
-)
-width_pfit_MR = (y_pis_pfit_MR[:, 1, 0] - y_pis_pfit_MR[:, 0, 0]).mean()
 
 # Print results
 print(
@@ -200,26 +165,15 @@ print(
     "\nEnbPI with partial_fit:"
     f"{coverage_pfit_enbpi:.3f}, {width_pfit_enbpi:.3f}"
 )
-print(
-    "Coverage / prediction interval width mean for MapieTimeSeriesRegressor: "
-    "\nPlus, with partial_fit:"
-    f"{coverage_pfit_plus:.3f}, {width_pfit_plus:.3f}"
-)
-print(
-    "Coverage / prediction interval width mean for MapieTimeSeriesRegressor: "
-    "\nMR_Like, with partial_fit:"
-    f"{coverage_pfit_MR:.3f}, {width_pfit_MR:.3f}"
-)
 
 # Plot estimated prediction intervals on test set
-fig, ((ax1, ax2), (ax3, ax4)) = plt.subplots(
-    nrows=2, ncols=2, figsize=(30, 25), sharey="row", sharex="col"
+fig, (ax1, ax2) = plt.subplots(
+    nrows=2, ncols=1, figsize=(30, 25), sharey="row", sharex="col"
 )
 
-for ax in [ax1, ax2, ax3, ax4]:
+for ax in [ax1, ax2]:
     ax.set_ylabel("Hourly demand (GW)")
     ax.plot(demand_test.Demand, lw=2, label="Test data", c="C1")
-
 
 ax1.plot(
     demand_test.index, y_pred_npfit_enbpi, lw=2, c="C2", label="Predictions"
@@ -253,39 +207,5 @@ ax2.set_title(
     f"Coverage:{coverage_pfit_enbpi:.3f}  Width:{width_pfit_enbpi:.3f}"
 )
 
-
-ax3.plot(
-    demand_test.index,
-    y_pred_pfit_plus,
-    lw=2,
-    c="C2",
-    label="Predictions",
-)
-ax3.fill_between(
-    demand_test.index,
-    y_pis_pfit_plus[:, 0, 0],
-    y_pis_pfit_plus[:, 1, 0],
-    color="C2",
-    alpha=0.2,
-    label="MapieTimeSeriesRegressor PIs",
-)
-ax3.set_title(
-    "Plus, with partial_fit.\n"
-    f"Coverage:{coverage_pfit_plus:.3f}"
-    f"Width:{width_pfit_plus:.3f}"
-)
-ax4.plot(demand_test.index, y_pred_pfit_MR, lw=2, c="C2", label="Predictions")
-ax4.fill_between(
-    demand_test.index,
-    y_pis_pfit_MR[:, 0, 0],
-    y_pis_pfit_MR[:, 1, 0],
-    color="C2",
-    alpha=0.2,
-    label="MapieTimeSeriesRegressor PIs",
-)
-ax4.set_title(
-    "MapieRegressor Like, with partial_fit\n"
-    f"Coverage:{coverage_pfit_MR:.3f}  Width:{width_pfit_MR:.3f}"
-)
 ax1.legend()
 plt.show()