+++ mapie/time_series_regression.py
@@ -73,69 +73,6 @@ class MapieTimeSeriesRegressor(MapieRegressor):
         y_pred, _ = super().predict(X, alpha=0.5, ensemble=True)
         return np.asarray(y) - np.asarray(y_pred)
 
-    def fit(
-        self,
-        X: ArrayLike,
-        y: ArrayLike,
-        sample_weight: Optional[ArrayLike] = None,
-    ) -> MapieTimeSeriesRegressor:
-        """
-        Compare to the method ``fit`` of ``MapieRegressor``, the ``fit`` method
-        of ``MapieTimeSeriesRegressor`` computes the ``conformity_scores_``
-        with relative values.
-
-        Returns
-        -------
-        MapieTimeSeriesRegressor
-            The model itself.
-        """
-        self = super().fit(X=X, y=y, sample_weight=sample_weight)
-        self.conformity_scores_ = self._relative_conformity_scores(X, y)
-        return self
-
-    def partial_fit(
-        self,
-        X: ArrayLike,
-        y: ArrayLike,
-    ) -> MapieTimeSeriesRegressor:
-        """
-        Update the ``conformity_scores_`` and ``k_`` attributes when new data
-        with known labels are available.
-        Note: Don't use ``partial_fit`` with samples of the training set.
-
-        Parameters
-        ----------
-        X : ArrayLike of shape (n_samples_test, n_features)
-            Input data.
-
-        y : ArrayLike of shape (n_samples_test,)
-            Input labels.
-
-        Returns
-        -------
-        MapieTimeSeriesRegressor
-            The model itself.
-
-        Raises
-        ------
-        ValueError
-            If the lenght of y is greater than the lenght of the training set.
-        """
-        X = cast(NDArray, X)
-        y = cast(NDArray, y)
-        if len(X) > len(self.conformity_scores_):
-            raise ValueError(
-                "You try to update more residuals than there are!"
-            )
-        new_conformity_scores_ = self._relative_conformity_scores(X, y)
-        self.conformity_scores_ = np.roll(
-            self.conformity_scores_, -len(new_conformity_scores_)
-        )
-        self.conformity_scores_[-len(new_conformity_scores_):] = (
-            new_conformity_scores_
-        )
-        return self
-
     def _beta_optimize(
         self,
         alpha: Union[float, NDArray],
@@ -143,8 +80,7 @@ class MapieTimeSeriesRegressor(MapieRegressor):
         lower_bounds: NDArray,
     ) -> NDArray:
         """
-        ``_beta_optimize`` offers to minimize the width of the PIs, for a given
-        difference of quantiles.
+        Minimize the width of the PIs, for a given difference of quantiles.
 
         Parameters
         ----------
@@ -172,7 +108,7 @@ class MapieTimeSeriesRegressor(MapieRegressor):
             )
         alpha = cast(NDArray, alpha)
         betas_0 = np.full(
-            shape=(len(alpha), len(lower_bounds)),
+            shape=(len(lower_bounds), len(alpha)),
             fill_value=np.nan,
             dtype=float,
         )
@@ -189,57 +125,96 @@ class MapieTimeSeriesRegressor(MapieRegressor):
                 1 - _alpha + betas,
                 axis=1,
                 method="higher",
-            )  # type: ignore
+            )
             beta = np_nanquantile(
                 lower_bounds,
                 betas,
                 axis=1,
                 method="lower",
-            )  # type: ignore
-            betas_0[ind_alpha, :] = betas[
+            )
+            betas_0[:, ind_alpha] = betas[
                 np.argmin(one_alpha_beta - beta, axis=0)
             ]
 
         return betas_0
 
-    def _pred_multi(self, X: ArrayLike) -> NDArray:
+    def fit(
+        self,
+        X: ArrayLike,
+        y: ArrayLike,
+        sample_weight: Optional[ArrayLike] = None,
+    ) -> MapieTimeSeriesRegressor:
         """
-        Return a prediction per train sample for each test sample, by
-        aggregation with matrix  ``k_``.
-
-        Parameters
-        ----------
-            X: NDArray of shape (n_samples_test, n_features)
-                Input data
+        Compared to the method ``fit`` of ``MapieRegressor``, the ``fit``
+        method of ``MapieTimeSeriesRegressor`` computes the
+        ``conformity_scores_`` with relative values.
 
         Returns
         -------
-            NDArray of shape (n_samples_test, n_samples_train)
+        MapieTimeSeriesRegressor
+            The model itself.
         """
-        y_pred_multi = np.column_stack(
-            [e.predict(X) for e in self.estimators_]
-        )
-        # At this point, y_pred_multi is of shape
-        # (n_samples_test, n_estimators_). The method
-        # ``aggregate_with_mask`` fits it to the right size
-        # thanks to the shape of k_.
+        self = super().fit(X=X, y=y, sample_weight=sample_weight)
+        self.conformity_scores_ = self._relative_conformity_scores(X, y)
+        return self
 
-        y_pred_multi = self.aggregate_with_mask(y_pred_multi, self.k_)
-        return y_pred_multi
+    def partial_fit(
+        self,
+        X: ArrayLike,
+        y: ArrayLike,
+    ) -> MapieTimeSeriesRegressor:
+        """
+        Update the ``conformity_scores_`` attribute when new data with known
+        labels are available.
+        Note: Don't use ``partial_fit`` with samples of the training set.
+
+        Parameters
+        ----------
+        X : ArrayLike of shape (n_samples_test, n_features)
+            Input data.
+
+        y : ArrayLike of shape (n_samples_test,)
+            Input labels.
+
+        Returns
+        -------
+        MapieTimeSeriesRegressor
+            The model itself.
+
+        Raises
+        ------
+        ValueError
+            If the length of y is greater than the length of the training set.
+        """
+        X = cast(NDArray, X)
+        y = cast(NDArray, y)
+        n = len(self.conformity_scores_)
+        if len(X) > n:
+            raise ValueError(
+                "You try to update more residuals than there are!"
+            )
+        new_conformity_scores_ = self._relative_conformity_scores(X, y)
+        self.conformity_scores_ = np.roll(
+            self.conformity_scores_, -len(new_conformity_scores_)
+        )
+        self.conformity_scores_[
+            -len(new_conformity_scores_) :
+        ] = new_conformity_scores_
+        return self
 
     def predict(
         self,
         X: ArrayLike,
         ensemble: bool = False,
         alpha: Optional[Union[float, Iterable[float]]] = None,
-        beta_optimize: bool = True,
+        optimize_beta: bool = True,
     ) -> Union[NDArray, Tuple[NDArray, NDArray]]:
         """
         Correspond to 'Conformal prediction for dynamic time-series'.
 
         Parameters
         ----------
-        beta_optimize: bool
+        optimize_beta: bool
             Whether to optimize the PIs' width or not.
 
         """
@@ -248,42 +223,36 @@ class MapieTimeSeriesRegressor(MapieRegressor):
         check_is_fitted(self, self.fit_attributes)
         self._check_ensemble(ensemble)
         alpha = cast(Optional[NDArray], check_alpha(alpha))
-        X = check_array(X, force_all_finite=False, dtype=["float64", "object"])
         y_pred = self.single_estimator_.predict(X)
+        n = len(self.conformity_scores_)
 
         if alpha is None:
             return np.array(y_pred)
         else:
             alpha_np = cast(NDArray, alpha)
-            check_alpha_and_n_samples(alpha_np, len(self.conformity_scores_))
+            check_alpha_and_n_samples(alpha_np, n)
 
-            if beta_optimize:
+            if optimize_beta:
                 betas_0 = self._beta_optimize(
-                    alpha=alpha_np,
-                    lower_bounds=self.conformity_scores_.reshape(1, -1),
-                    upper_bounds=self.conformity_scores_.reshape(1, -1),
+                    alpha_np,
+                    self.conformity_scores_.reshape(1, -1),
+                    self.conformity_scores_.reshape(1, -1),
                 )
             else:
-                betas_0 = np.full(
-                    shape=(len(alpha), len(self.conformity_scores_)),
-                    fill_value=np.nan,
-                    dtype=float,
-                )
-                for ind_alpha, _alpha in enumerate(alpha):
-                    betas_0[ind_alpha, :] = _alpha / 2.0
+                betas_0 = np.repeat(alpha[:, np.newaxis] / 2, n, axis=0)
 
             lower_quantiles = np_nanquantile(
                 self.conformity_scores_,
-                betas_0[:, 0],
+                betas_0[0, :],
                 axis=0,
                 method="lower",
-            ).T  # type: ignore
+            ).T
             higher_quantiles = np_nanquantile(
                 self.conformity_scores_,
-                1 - alpha_np + betas_0[:, 0],
+                1 - alpha_np + betas_0[0, :],
                 axis=0,
                 method="higher",
-            ).T  # type: ignore
+            ).T
 
             if self.cv == "prefit":
                 y_pred_low = y_pred[:, np.newaxis] + lower_quantiles
@@ -293,18 +262,8 @@ class MapieTimeSeriesRegressor(MapieRegressor):
                 pred = aggregate_all(self.agg_function, y_pred_multi)
                 lower_bounds, upper_bounds = pred, pred
 
-                y_pred_low = np.column_stack(
-                    [
-                        lower_bounds + lower_quantiles[k]
-                        for k, _ in enumerate(alpha_np)
-                    ]
-                )
-                y_pred_up = np.column_stack(
-                    [
-                        upper_bounds + higher_quantiles[k]
-                        for k, _ in enumerate(alpha_np)
-                    ]
-                )
+                y_pred_low = lower_bounds + lower_quantiles.reshape(-1, 1)
+                y_pred_up = upper_bounds + higher_quantiles.reshape(-1, 1)
 
                 if ensemble:
                     y_pred = aggregate_all(self.agg_function, y_pred_multi)