+++ HISTORY.rst
@@ -5,6 +5,7 @@ History
 0.3.3 (2022-XX-XX)
 ------------------
 * Relax and fix typing
+* Add EnbPI method for Time Series Regression
 
 0.3.2 (2022-03-11)
 ------------------
@@ -17,7 +18,6 @@ History
 * Uniformize the use of matrix k_ and add an argument "ensemble" to method "predict" in regression.py
 * Add replication of the Chen Xu's tutorial testing Jackknife+aB vs Jackknife+
 * Add Jackknife+-after-Bootstrap documentation
-* Add EnbPI method for Time Series Regression
 * Improve scikit-learn pipelines compatibility
 
 0.3.1 (2021-11-19)